增加自动回复功能 1. 语音识别错误时,触发自动回复。 2. 语音识别文字长度异常, 触发自动回复。 新增固定回复模块 可以自定义特殊情况时的自定义回复语音内容

+++ funcs.py
@@ -155,13 +155,14 @@ async def ask_model(prompt: list, lang_type='zh'):
             break
 
 
-async def send_audio_data(audio_path, text_response, history, status):
+async def send_audio_data(audio_path, text_response, history, status, delete=True):
     async with aiofiles.open(audio_path, 'rb') as f:
         temp_data = await f.read()
 
     # todo 这里将该音频文件删除一下
     try:
-        os.remove(audio_path)
+        if delete:
+            os.remove(audio_path)
     except Exception:
         pass
